+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- close modal panels on return key in model panel entry fields on non Android and non iOS.
 
 ## [1.7.0] - 2025-12-25